microsoft_botframework/requirements.txt
botbuilder-core>=4.4.0b1
coco-puppet
